--- src/PropertyNameHelper.php
@@ -1,9 +1,0 @@
-<?php
-
-
-namespace Enz0project\ModelDocumenter;
-
-
-interface PropertyNameHelper {
-	public function propertyNameFromColumn(string $column): string;
-}
